sidemenu as component
